requirements.txt
streamlit
opencv-python-headless
numpy
tensorflow-cpu
streamlit-webrtc